+++ cmd/dump.go
@@ -9,9 +9,10 @@ import (
 func newDumpCmd(opts *lefthook.Options) *cobra.Command {
 	dumpArgs := lefthook.DumpArgs{}
 	dumpCmd := cobra.Command{
-		Use:     "dump",
-		Short:   "Prints config merged from all extensions (in YAML format by default)",
-		Example: "lefthook dump",
+		Use:               "dump",
+		Short:             "Prints config merged from all extensions (in YAML format by default)",
+		Example:           "lefthook dump",
+		ValidArgsFunction: cobra.NoFileCompletions,
 		Run: func(cmd *cobra.Command, args []string) {
 			lefthook.Dump(opts, dumpArgs)
 		},