+++ internal/log/skip_settings.go
@@ -1,5 +1,9 @@
 package log
 
+import (
+	"strings"
+)
+
 const (
 	skipMeta = 1 << iota
 	skipSuccess
@@ -10,11 +14,28 @@ const (
 	skipExecutionOutput
 	skipExecutionInfo
 	skipEmptySummary
+	skipAll = (1 << iota) - 1
 )
 
 type SkipSettings int16
 
-func (s *SkipSettings) ApplySetting(setting string) {
+func (s *SkipSettings) ApplySettings(tags string, skipOutput interface{}) {
+	switch typedSkipOutput := skipOutput.(type) {
+	case bool:
+		s.SkipAll(typedSkipOutput)
+	case []interface{}:
+		for _, skipOption := range typedSkipOutput {
+			s.applySetting(skipOption.(string))
+		}
+		if tags != "" {
+			for _, skipOption := range strings.Split(tags, ",") {
+				s.applySetting(skipOption)
+			}
+		}
+	}
+}
+
+func (s *SkipSettings) applySetting(setting string) {
 	switch setting {
 	case "meta":
 		*s |= skipMeta
@@ -34,6 +55,14 @@ func (s *SkipSettings) ApplySetting(setting string) {
 		*s |= skipExecutionInfo
 	case "empty_summary":
 		*s |= skipEmptySummary
+	}
+}
+
+func (s *SkipSettings) SkipAll(val bool) {
+	if val {
+		*s = skipAll &^ skipFailure
+	} else {
+		*s = 0
 	}
 }