+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## master (unreleased)
 
+## 1.5.7 (2024-01-17)
+
+- fix: pre push hook handling ([#613](https://github.com/evilmartians/lefthook/pull/613)) by @mrexox
+- chore: hide wiki links ([#608](https://github.com/evilmartians/lefthook/pull/608)) by @mrexox
+
+## 1.5.6 (2024-01-12)
+
+- feat: shell completion improvements ([#577](https://github.com/evilmartians/lefthook/pull/577)) by @scop
+- fix: safe execute git commands without sh wrapper ([#606](https://github.com/evilmartians/lefthook/pull/606)) by @mrexox
+- fix: use lefthook package with npx ([#604](https://github.com/evilmartians/lefthook/pull/604)) by @mrexox
+- feat: allow setting a bool value for skip_output ([#601](https://github.com/evilmartians/lefthook/pull/601)) by @nsklyarov
+- docs: update exception case about interactive option by @mrexox
+
 ## 1.5.5 (2023-11-30)
 
 - fix: use empty stdin by default ([#590](https://github.com/evilmartians/lefthook/pull/590)) by @mrexox